+++ packages/orq-rc/RELEASES.md
@@ -9399,3 +9399,13 @@ Based on:
 - [typescript v4.10.0-rc.26] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.26] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.26 - packages/orq-rc
+
+## 2026-05-22 06:05:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.764.0 (2.885.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.27] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.27] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.27 - packages/orq-rc